requirements.txt
# CRC Portal — Dependencies
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Mail>=0.10
pandas>=2.0
openpyxl>=3.1
gspread>=5.10
google-auth>=2.23
python-dotenv>=1.0
Werkzeug>=3.0
mcp>=1.0
